Prettyprinter.Render.Terminal

Description

Render SimpleDocStream in a terminal.

Styling

data AnsiStyle

Render the annotated document in a certain style. Styles not set in the annotation will use the style of the surrounding document, or the terminal’s default if none has been set yet.

style = color Green <> bold
styledDoc = annotate style "hello world"

data Color

The 8 ANSI terminal colors.

Constructors

Black

Red

Green

Yellow

Blue

Magenta

Cyan

White

Font color

color :: Color -> AnsiStyle

Style the foreground with a vivid color.

colorDull :: Color -> AnsiStyle

Style the foreground with a dull color.

Background color

bgColor :: Color -> AnsiStyle

Style the background with a vivid color.

bgColorDull :: Color -> AnsiStyle

Style the background with a dull color.

Font style

bold :: AnsiStyle

Render in bold.

italicized :: AnsiStyle

Render in italics.

underlined :: AnsiStyle

Render underlined.

Internal markers

These should only be used for writing adaptors to other libraries; for the average use case, use bold, bgColorDull, etc.

data Intensity

Dull or vivid coloring, as supported by ANSI terminals.

Constructors

Vivid

Dull

data Bold

Constructors

Bold

data Underlined

Constructors

Underlined

data Italicized

Constructors

Italicized

Conversion to ANSI-infused Text

renderLazy :: SimpleDocStream AnsiStyle -> Text

(renderLazy doc) takes the output doc from a rendering function and transforms it to lazy text, including ANSI styling directives for things like colorization.

ANSI color information will be discarded by this function unless you are running on a Unix-like operating system. This is due to a technical limitation in Windows ANSI support.

With a bit of trickery to make the ANSI codes printable, here is an example that would render colored in an ANSI terminal:

>>> let render = TL.putStrLn . TL.replace "\ESC" "\\e" . renderLazy . layoutPretty defaultLayoutOptions
>>> let doc = annotate (color Red) ("red" <+> align (vsep [annotate (color Blue <> underlined) ("blue+u" <+> annotate bold "bold" <+> "blue+u"), "red"]))
>>> render (unAnnotate doc)
red blue+u bold blue+u
    red
>>> render doc
\e[0;91mred \e[0;94;4mblue+u \e[0;94;1;4mbold\e[0;94;4m blue+u\e[0;91m
    red\e[0m

Run the above via echo -e ... in your terminal to see the coloring.

renderStrict :: SimpleDocStream AnsiStyle -> Text

(renderStrict sdoc) takes the output sdoc from a rendering and transforms it to strict text.

Render directly to stdout

renderIO :: Handle -> SimpleDocStream AnsiStyle -> IO ()

(renderIO h sdoc) writes sdoc to the handle h.

>>> let render = renderIO System.IO.stdout . layoutPretty defaultLayoutOptions
>>> let doc = annotate (color Red) ("red" <+> align (vsep [annotate (color Blue <> underlined) ("blue+u" <+> annotate bold "bold" <+> "blue+u"), "red"]))

We render the unAnnotated version here, since the ANSI codes don’t display well in Haddock,

>>> render (unAnnotate doc)
red blue+u bold blue+u
    red

This function behaves just like

renderIO h sdoc = hPutStr h (renderLazy sdoc)

but will not generate any intermediate text, rendering directly to the handle.

Convenience functions

putDoc :: Doc AnsiStyle -> IO ()

(putDoc doc) prettyprints document doc to standard output using defaultLayoutOptions.

>>> putDoc ("hello" <+> "world")
hello world

putDoc = hPutDoc stdout

hPutDoc :: Handle -> Doc AnsiStyle -> IO ()

Like putDoc, but instead of using stdout, print to a user-provided handle, e.g. a file or a socket using defaultLayoutOptions.

main = withFile "someFile.txt" (\h -> hPutDoc h (vcat ["vertical", "text"]))

hPutDoc h doc = renderIO h (layoutPretty defaultLayoutOptions doc)